requirements.txt
python-dotenv==1.0.0
requests==2.31.0
groq==1.1.0
customtkinter==5.2.0
Pillow==10.1.0
BlurWindow==1.2.1
keyboard==0.13.5
pystray==0.19.5
pywin32==306
pyautogui==0.9.54
pywhatkit==5.4
pygetwindow==0.0.9
psutil==5.9.6
pyperclip==1.8.2
pynput==1.7.6
pyttsx3==2.90
SpeechRecognition==3.10.0
edge-tts==7.2.7
opencv-python==4.13.0.92
numpy==2.4.3
schedule==1.2.0
langdetect==1.0.9
cryptography==41.0.7
pygame==2.6.1
screen-brightness-control==0.27.0
Send2Trash==2.1.0
wikipedia==1.4.0
python-telegram-bot==22.6
sounddevice==0.5.5
plyer==2.1.0